Rename SnykBuildStep to SnykStepBuilder
we use this class for freestyle and pipeline jobs, so no
separation via packages
