+++ src/main/java/io/snyk/jenkins/Severity.java
@@ -1,4 +1,4 @@
-package io.snyk.jenkins.steps;
+package io.snyk.jenkins;
 
 import java.util.stream.Stream;
