+++ src/main/resources/io/snyk/jenkins/SnykStepBuilder/config.jelly
